Move common response models to new package
